microservice1/requirements.txt
nats-py==2.8.0
aiohttp==3.8.1
aiohttp-cors==0.7.0
async-timeout==4.0.0
asyncio-nats-client==0.11.2
colorclass==2.2.0
requests==2.24.0
six==1.15.0